+++ MicrobiologicalExamination.java
@@ -2,7 +2,7 @@ public class MicrobiologicalExamination extends Exam {
     private String sampleType;
     final double costIncreaseRate = 0.2;
 
-    public MicrobiologicalExamination(int examID, String examName, String examCategory, int maxSlotsPerDay, int cost, int doctorID, String sampleType){
+    public MicrobiologicalExamination(int examID, String examName, String examCategory, int maxSlotsPerDay, double cost, int doctorID, String sampleType){
         super(examID, examName, examCategory, maxSlotsPerDay, cost, doctorID);
         this.sampleType=sampleType;
     }
